+++ stories/api/index.ts
@@ -5,3 +5,4 @@ import './Properties';
 import './Requirements';
 import './Item';
 import './ExtendedStats';
+import './Body';